--- src/app/components/ProductsCardInventoryList.tsx
@@ -1,11 +1,0 @@
-import products from '@/data/dummyproducts.json';
-
-import ProductsCard from "./ProductsCard";
-
-export default function ProductsCardInventoryList() {
-  return (
-    <ul className="flex [&>li]:border [&>li]:p-4 gap-4">
-      <ProductsCard maxProducts={4} products={products} hasLike/>
-    </ul>
-  )
-}
